All activites, udps and actions are mostly done. Some rvp stuff done. These are all the classes we need I beleive.

+++ src/simModel/CastingStation.java
@@ -5,6 +5,5 @@ public class CastingStation {
     public Bin bin;
     public int type;
     public int status;
-    public int timeToNextBreak;
-
+    public double timeToNextBreak;
 }
